Some proposed API changes.
The main one is to make the dungeon infos available via a dictionary. This makes a lot more sense to me as it means that adding a new dungeon requires only adding a new value to the DungeonKey enum. I think it also makes them easier to work with as the dungeons naturally become iterable.

Also fixed some bugs that I found.